# source/bot/components/participant_drawer/src/openapi_server/models/draw_result.py
# coding: utf-8

"""
    Topics and Draw results API

    Этот API предназначен для управления темами мероприятий и работы с жеребьёвкой. Позволяет создавать, редактировать и удалять темы, а также запускать и получать распределение тем между группами участников. 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field
from typing import Any, ClassVar, Dict, List
from typing_extensions import Annotated
from openapi_server.models.draw_result_topics_inner import DrawResultTopicsInner
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class DrawResult(BaseModel):
    """
    Распределение тем по группам
    """ # noqa: E501
    group_id: Annotated[int, Field(strict=True, ge=1)]
    group_name: Annotated[str, Field(min_length=1, strict=True)]
    topics: List[DrawResultTopicsInner] = Field(description="Список тем, назначенных группе (обычно 1)")
    __properties: ClassVar[List[str]] = ["group_id", "group_name", "topics"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of DrawResult from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of each item in topics (list)
        _items = []
        if self.topics:
            for _item in self.topics:
                if _item:
                    _items.append(_item.to_dict())
            _dict['topics'] = _items
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of DrawResult from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "group_id": obj.get("group_id"),
            "group_name": obj.get("group_name"),
            "topics": [DrawResultTopicsInner.from_dict(_item) for _item in obj.get("topics")] if obj.get("topics") is not None else None
        })
        return _obj